# api-contract/src/main/resources/static/brightlife_api_contract.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Bright-life Enhancement API
  description:
    "API documentation for Bright-life Enhancement project"
  license:
    name: "fronteersdev Inc"
    url: "https://fronteersdev.com/license"
servers:
  - url: http://localhost:7081
    description: Local test server
  - url: https://bright-life-enhancement.fronteers-dev.com/
    description: Dev test server
security:
  - BearerAuth: [ ]  # ⬅️ Apply JWT security globally to all endpoints
components:
  securitySchemes:
    BearerAuth:
      type: http
      description: This API uses JWT for authentication
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      explode: true
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      style: form
    limitParam:
      explode: true
      in: query
      name: limit
      required: false
      schema:
        format: int32
        maximum: 100
        type: integer
      style: form
...
    idParam:
      in: path
      name: idParam
      description: A unique id to update or fetch a record.
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    PaginatedReviews:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        itemsPerPage:
          type: integer
          example: 100
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        email:
          type: string
        referralWish:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewMessage:
          type: string
        status:
          $ref: '#/components/schemas/ReviewStatusEnum'
    ReviewStatusEnum:
      type: string
      enum:
        - draft
        - published
...
    ReviewSearch:
      type: object
      properties:
        nickName:
          type: string
        email:
          $ref: "#/components/schemas/Email"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        status:
          $ref: '#/components/schemas/ReviewStatusEnum'
...
    Email:
      type: string
      #      pattern: "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+${5,200}"
      format: email
...
    Error:
      description: |

        **The below table describes custom error codes returned by various APIs currently**
        | Code      | error         | Message                     |
        | ----------| --------------|-----------------------------|
        | 400       | INPUT         | BAD REQUEST                 |
        | 404       | NOT_FOUND     | RESOURCE DOES NOT EXIST     |
        | 422       | UNPROCESSABLE | UNPROCESSABLE ENTITY        |
        | 401       | UNAUTHORIZED  | UNAUTHORIZED                |
        | 403       | FORBIDDEN     | FORBIDDEN                   |
        | 500       | PROCESSING    | INTERNAL SERVER ERROR       |
        | 500       | TIMEOUT       | TIMEOUT                     |
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
    Success:
      type: object
      required:
        - status
        - success
        - message
      properties:
        status:
          type: boolean
          default: false
        success:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      #      type: object
      description: |
        Bad Request
        The request body or parameters does not conform to the schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        Unauthorized.

        - Indicates the request does not contain a valid token to authenticate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Forbidden.
            - Indicates the authenticated entity does not have the privilege to perform
              the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnProcessableEntity:
      description: |
        Unprocessable entity.
        - The request body may be well-formed, but either request payload or parameters failed semantic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: |
        Internal Server Error.
            - The server encountered an unexpected error while processing the request..
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /api/v1/reviews:
    post:
      summary: Submits A Review
      tags:
        - Review
      operationId: submitReview
      requestBody:
        description: Patients have the opportunity to share their personal experiences and testimonials with others, providing a platform for them to express their thoughts, feelings, and outcomes. This allows other patients to gain insight and perspective from those who have gone through similar experiences, fostering a sense of community and support.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
      responses:
        "201":
          description: Review Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/search:
    post:
      summary:
        Searches and returns all Reviews matching all of the search criteria fields passed.
        The number of search results shall be constrained by the limit params in Query.
        Search criteria, applies an **AND** based search on all fields passed in the request.
        If the search criteria is empty/null, it returns list of reviews.
      tags:
        - Review
      operationId: listReviews
      parameters:
        - description: The page number (starting from 1)
          $ref: "#/components/parameters/PageParam"
        - description: How many items to return at one time (max 100)
          $ref: "#/components/parameters/limitParam"
      requestBody:
        description:
          "Search criteria, applies an **AND** based search on all fields\
                \ passed in the request."
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSearch"
      responses:
        "200":
          description: Review Records Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/{idParam}:
    get:
      summary: Fetches A Review Record Using A Uniques idParam
      tags:
        - Review
      operationId: getReview
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Review Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/publish/{idParam}:
    put:
      summary: Fetches A Review Record Using A Unique idParam And Publishes it
      tags:
        - Review
      operationId: publish
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Review Published Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/unpublish/{idParam}:
    put:
      summary: Fetches A Review Record Using A Unique idParam And Un-publishes it, thereby Sending It To The List Of Drafts
      tags:
        - Review
      operationId: unpublish
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Review Unpublished Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/published:
    get:
      summary: Fetches All Published Reviews
      tags:
        - Review
      operationId: getPublished
      parameters:
        - description: The page number (starting from 1)
          $ref: "#/components/parameters/PageParam"
        - description: How many items to return at one time (max 100)
          $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: Published Reviews Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reviews/drafts:
    get:
      summary: Fetches All Draft Reviews
      tags:
        - Review
      operationId: getDrafts
      parameters:
        - description: The page number (starting from 1)
          $ref: "#/components/parameters/PageParam"
        - description: How many items to return at one time (max 100)
          $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: Draft Reviews Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...